refactor: move rust codegen into separate module
