+++ src/lib.rs
@@ -1,10 +1,11 @@
 #![doc = include_str!("../README.md")]
 
-#[cfg(feature = "codegen")]
-pub use codegen::configure;
 pub use parser::parse;
 
 #[cfg(feature = "codegen")]
-mod codegen;
+pub use codegen::rust::configure;
+
+#[cfg(feature = "codegen")]
+pub mod codegen;
 pub(crate) mod parser;
 pub mod schema;
